+++ src/LIVI_Depth_model_pi_UI_impl.h
@@ -77,12 +77,12 @@ public:
 
     void        SetCustomColor(int num, wxColour& col);
     wxColour    GetCustomColor(int num);
-    void        SetCustomLevel(int num, int level);
+    void        SetCustomLevel(int num, double level);
     double      GetCustomLevel(int num);
 
     void        SetTwoColours(int num, wxColour& col);
     wxColour    GetTwoColours(int num);
-    void        SetDividingLevel(int level);
+    void        SetDividingLevel(double level);
     double      GetDividingLevel();
 
     void        SetCurrentWaterLevel(double level);
